Smol fixes/optimizations (#65)
* Fixes code-423n4/2022-02-hubble-findings#130

* MINTER_ROLE to deployer only in RestrictedVusd

* cleanup txOptions

* Fixes code-423n4/2022-02-hubble-findings#119

* ch loop optimiz

* uint i;

* emit FundingPaid

+++ contracts/Interfaces.sol
@@ -33,7 +33,7 @@ interface IClearingHouse {
     function updatePositions(address trader) external;
     function getMarginFraction(address trader) external view returns(int256);
     function getTotalFunding(address trader) external view returns(int256 totalFunding);
-    function getAmmsLength() external view returns(uint8);
+    function getAmmsLength() external view returns(uint);
     function amms(uint idx) external view returns(IAMM);
     function maintenanceMargin() external view returns(int256);
     function minAllowableMargin() external view returns(int256);
